# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    [OpenID Spec](https://openid.net/specs/openid-sse-framework-1_0.html#management)

    HTTP API to be implemented by Event Transmitters. This API can be used by
    Event Receivers to query and update the Event Stream configuration and
    status, to add and remove subjects, and to trigger verification.
  license:
    name: MIT
  title: Stream Management API for OpenID Shared Security Events
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /status:
    get:
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#reading-a-streams-status)

        An Event Receiver checks the current status of an event stream by making an
        HTTP GET request to the stream’s Status Endpoint.
      operationId: get_status
      parameters:
      - description: |-
          OPTIONAL. The subject for which the stream status is requested.
          Content is a string-encoded json representation of a Subject
          (see Subject schema)
        explode: true
        in: query
        name: subject
        required: false
        schema:
          example: "{\"user\":{\"format\":\"email\",\"email\":\"reginold@popular-app.com\"\
            }}"
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StreamStatus'
          description: |-
            On receiving a valid request the Event Transmitter responds with a 200
            OK response containing a JSON object with an attribute status
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Authorization failed or is missing
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The Event Receiver is not allowed to read this stream's status.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: |-
            There is no Event Stream configured for this Event Receiver,
            or the Subject specified is invalid or
            the Receiver is not authorized to get status for the specified Subject.
      security:
      - BearerAuth: []
      summary: Request to get the status of an Event Stream
      tags:
      - StreamManagement
      x-accepts: application/json
      x-tags:
      - tag: StreamManagement
    post:
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0-ID1.html#updating-a-streams-status)

        An Event Receiver updates the current status of a stream by making an HTTP
        POST request to the Status Endpoint.
      operationId: update_status
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStreamStatus'
        description: Request parameters
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateStreamStatus'
          description: Event Stream status successfully updated
        "202":
          description: |-
            The update request has been accepted, but not processed.
            Receiver MAY try the same request later in order to get processing result.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Request body cannot be parsed or the request is otherwise invalid
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Authorization failed or is missing
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The Event Receiver is not allowed to update the stream status.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: |-
            There is no Event Stream configured for this Event Receiver,
            or an invalid Subject is specified.
      security:
      - BearerAuth: []
      summary: Request to update an Event Stream's status
      tags:
      - StreamManagement
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: StreamManagement
...
  /register:
    post:
      description: |-
        This endpoint is not part of the spec, but rather a quick-and-dirty way to allow a receiver to pass along its
        audience claim in exchange for a bearer token.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterParameters'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
          description: "On successful creation of a stream, returns a bearer token\
            \ that will be used for subsequent requests."
      summary: Request the transmitter to setup a stream for a given receiver.
      tags:
      - OutOfBand
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: OutOfBand
...
components:
...
  schemas:
    StreamStatus:
      example:
        subject:
          format: email
          email: reginold@popular-app.com
        status: disabled
      properties:
        status:
          description: |-
            REQUIRED. The status of the stream. Values can be one of:

            enabled:
              The Transmitter MUST transmit events over the stream,
              according to the stream’s configured delivery method.

            paused:
              The Transmitter MUST NOT transmit events over the stream.
              The transmitter will hold any events it would have transmitted while paused,
              and SHOULD transmit them when the stream’s status becomes enabled.
              If a Transmitter holds successive events that affect the same Subject Principal,
              then the Transmitter MUST make sure that those events are transmitted in
              the order of time that they were generated OR the Transmitter MUST send
              only the last events that do not require the previous events affecting
              the same Subject Principal to be processed by the Receiver,
              because the previous events are either cancelled by the later events or
              the previous events are outdated.

            disabled:
              The Transmitter MUST NOT transmit events over the stream,
              and will not hold any events for later transmission.
          example: disabled
          type: string
        subject:
          allOf:
          - $ref: '#/components/schemas/Subject'
          description: OPTIONAL. The Subject to which the status applies.
          example:
            format: email
            email: reginold@popular-app.com
      required:
      - status
      type: object
...
    UpdateStreamStatus:
      allOf:
      - $ref: '#/components/schemas/StreamStatus'
      example:
        reason: Disabled by administrator action.
      properties:
        reason:
          description: OPTIONAL. A short text description that explains the reason
            for the change.
          example: Disabled by administrator action.
          type: string
      type: object
...
    RegisterResponse:
      example:
        token: 49e5e7785e4e4f688aa49e2585970370
      properties:
        token:
          description: The Bearer Token that the receiver will need to use for all
            Stream Management API calls that require authorization.
          type: string
      required:
      - token
      type: object
...
    RegisterParameters:
      example:
        audience: https://popular-app.com
      properties:
        audience:
          description: The audience claim to be used for all events on this stream.
          example: https://popular-app.com
          format: uri
          type: string
      type: object
...
    Account:
      description: |-
        [Spec](https://datatracker.ietf.org/doc/html/draft-ietf-secevent-subject-identifiers#section-3.2.1)

        The Account Identifier Format identifies a subject using an account
        at a service provider, identified with an "acct" URI as defined in
        [RFC7565](https://datatracker.ietf.org/doc/html/rfc7565).
        Subject Identifiers in this format MUST contain a "uri"
        member whose value is the "acct" URI for the subject.  The "uri"
        member is REQUIRED and MUST NOT be null or empty.  The Account
        Identifier Format is identified by the name "account".
      properties:
        format:
          default: account
          title: Format
          type: string
        uri:
          example: acct:reginold@popular-app.com
          pattern: "^acct:[^\\s]+$"
          title: Uri
          type: string
      required:
      - format
      - uri
      title: Account
      type: object
    DID:
      description: |-
        [Spec](https://datatracker.ietf.org/doc/html/draft-ietf-secevent-subject-identifiers#section-3.2.3)

        The Decentralized Identifier Format identifies a subject using a
        Decentralized Identifier (DID) URL as defined in [DID](https://www.w3.org/TR/did-core/).
        Subject Identifiers in this format MUST contain a "url" member whose value is
        a DID URL for the DID Subject being identified.  The value of the
        "url" member MUST be a valid DID URL and MAY be a bare DID.  The
        "url" member is REQUIRED and MUST NOT be null or empty.  The
        Decentralized Identifier Format is identified by the name "did".
      properties:
        format:
          default: did
          title: Format
          type: string
        url:
          example: did:example:123456/did/url/path?versionId=1
          pattern: "^did:[^\\s]+$"
          title: Url
          type: string
      required:
      - format
      - url
      title: DID
      type: object
    Email:
      description: |-
        [Spec](https://datatracker.ietf.org/doc/html/draft-ietf-secevent-subject-identifiers#section-3.2.4)

        The Email Identifier Format identifies a subject using an email
        address.  Subject Identifiers in this format MUST contain an "email"
        member whose value is a string containing the email address of the
        subject, formatted as an "addr-spec" as defined in Section 3.4.1 of
        [RFC5322](https://datatracker.ietf.org/doc/html/rfc5322).
        The "email" member is REQUIRED and MUST NOT be null or
        empty.  The value of the "email" member SHOULD identify a mailbox to
        which email may be delivered, in accordance with [RFC5321](https://datatracker.ietf.org/doc/html/rfc5321).
        The Email Identifier Format is identified by the name "email".
      properties:
        format:
          default: email
          title: Format
          type: string
        email:
          example: reginold@popular-app.com
          pattern: "(?:[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+/=?^_`{|}~-]+)*|\"\
            (?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\\\
            [\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\\
            .)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9]))\\\
            .){3}(?:(2(5[0-5]|[0-4][0-9])|1[0-9][0-9]|[1-9]?[0-9])|[a-z0-9-]*[a-z0-9]:(?:[\\\
            x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\\\[\\x01-\\x09\\\
            x0b\\x0c\\x0e-\\x7f])+)\\])"
          title: Email
          type: string
      required:
      - email
      - format
      title: Email
      type: object
    IssSub:
      description: |-
        [Spec](https://datatracker.ietf.org/doc/html/draft-ietf-secevent-subject-identifiers#section-3.2.5)

        The Issuer and Subject Identifier Format identifies a subject using a
        pair of "iss" and "sub" members, analagous to how subjects are
        identified using the "iss" and "sub" claims in OpenID Connect
        [OpenID.Core](http://openid.net/specs/openid-connect-core-1_0.html) ID Tokens.
        These members MUST follow the formats of the "iss" member and "sub" member
        defined by [RFC7519](https://datatracker.ietf.org/doc/html/rfc7519), respectively.
        Both the "iss" member and the "sub" member are REQUIRED and MUST NOT
        be null or empty.  The Issuer and Subject Identifier Format is
        identified by the name "iss_sub".
      properties:
        format:
          default: iss_sub
          title: Format
          type: string
        iss:
          example: https://most-secure.com
          title: iss
          type: string
        sub:
          example: "145234573"
          title: sub
          type: string
      required:
      - format
      - iss
      - sub
      title: IssSub
      type: object
    JwtID:
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#rfc.section.3.4.1)

        The "JWT ID" Subject Identifier Format specifies a JSON Web Token (JWT)
        identifier, defined in [RFC7519](https://datatracker.ietf.org/doc/html/rfc7519).
      properties:
        format:
          default: jwt_id
          title: Format
          type: string
        iss:
          description: |-
            The "iss" (issuer) claim of the JWT being identified, defined in
            [RFC7519](https://datatracker.ietf.org/doc/html/rfc7519)
          example: https://most-secure.com
          title: iss
          type: string
        jti:
          description: |-
            The "jti" (JWT token ID) claim of the JWT being identified, defined in
            [RFC7519](https://datatracker.ietf.org/doc/html/rfc7519)
          example: B70BA622-9515-4353-A866-823539EECBC8
          title: jti
          type: string
      required:
      - format
      - iss
      - jti
      title: JwtID
      type: object
    Opaque:
      description: |-
        [Spec](https://datatracker.ietf.org/doc/html/draft-ietf-secevent-subject-identifiers#section-3.2.6)

        The Opaque Identifier Format describes a subject that is identified
        with a string with no semantics asserted beyond its usage as an
        identifier for the subject, such as a UUID or hash used as a
        surrogate identifier for a record in a database.  Subject Identifiers
        in this format MUST contain an "id" member whose value is a JSON
        string containing the opaque string identifier for the subject.  The
        "id" member is REQUIRED and MUST NOT be null or empty.  The Opaque
        Identifier Format is identified by the name "opaque".
      properties:
        format:
          default: opaque
          title: Format
          type: string
        id:
          example: "11112222333344445555"
          title: Id
          type: string
      required:
      - format
      - id
      title: Opaque
      type: object
...
    SimpleSubject:
      discriminator:
        propertyName: format
      oneOf:
      - $ref: '#/components/schemas/Account'
      - $ref: '#/components/schemas/DID'
      - $ref: '#/components/schemas/Email'
      - $ref: '#/components/schemas/IssSub'
      - $ref: '#/components/schemas/JwtID'
      - $ref: '#/components/schemas/Opaque'
      - $ref: '#/components/schemas/PhoneNumber'
      - $ref: '#/components/schemas/SamlAssertionID'
      title: Simple Subject
      x-one-of-name: SimpleSubject
    Aliases:
      description: |-
        [Spec](https://datatracker.ietf.org/doc/html/draft-ietf-secevent-subject-identifiers#section-3.2.2)

        The Aliases Identifier Format describes a subject that is identified
        with a list of different Subject Identifiers.  It is intended for use
        when a variety of identifiers have been shared with the party that
        will be interpreting the Subject Identifier, and it is unknown which
        of those identifiers they will recognize or support.  Subject
        Identifiers in this format MUST contain an "identifiers" member whose
        value is a JSON array containing one or more Subject Identifiers.
        Each Subject Identifier in the array MUST identify the same entity.
        The "identifiers" member is REQUIRED and MUST NOT be null or empty.
        It MAY contain multiple instances of the same Identifier Format
        (e.g., multiple Email Subject Identifiers), but SHOULD NOT contain
        exact duplicates.  This format is identified by the name "aliases".

        "alias" Subject Identifiers MUST NOT be nested; i.e., the
        "identifiers" member of an "alias" Subject Identifier MUST NOT
        contain a Subject Identifier in the "aliases" format.'
      properties:
        format:
          default: aliases
          title: Format
          type: string
        identifiers:
          items:
            $ref: '#/components/schemas/SimpleSubject'
          minItems: 1
          type: array
      required:
      - format
      - identifiers
      title: Aliases
      type: object
    ComplexSubject:
      additionalProperties: false
      description: |-
        [Spec](https://openid.net/specs/openid-sse-framework-1_0.html#rfc.section.3.2)

        A Complex Subject Member has a name and a value that is a JSON object that
        has one or more Simple Subject Members. All members within a Complex Subject
        MUST represent attributes of the same Subject Principal.
        As a whole, the Complex Subject MUST refer to exactly one Subject Principal.
      minProperties: 1
      properties:
        application:
          $ref: '#/components/schemas/ComplexSubject_application'
        device:
          $ref: '#/components/schemas/ComplexSubject_device'
        group:
          $ref: '#/components/schemas/ComplexSubject_group'
        org_unit:
          $ref: '#/components/schemas/ComplexSubject_org_unit'
        session:
          $ref: '#/components/schemas/ComplexSubject_session'
        tenant:
          $ref: '#/components/schemas/ComplexSubject_tenant'
        user:
          $ref: '#/components/schemas/ComplexSubject_user'
      title: ComplexSubject
      type: object
    Subject:
      description: |-
        [Spec](https://datatracker.ietf.org/doc/html/draft-ietf-secevent-subject-identifiers#section-3)

        As described in Section 1.2 of SET [RFC8417](https://datatracker.ietf.org/doc/html/rfc8417),
        subjects related to security events may take a variety of forms,
        including but not limited to a JWT [RFC7519](https://datatracker.ietf.org/doc/html/rfc7519)
        principal, an IP address, a URL, etc. Different types of subjects may need
        to be identified in different ways. (e.g., a host might be identified by an
        IP or MAC address, while a user might be identified by an email address)
        Furthermore, even in the case where the type of the subject is known,
        there may be multiple ways by which a given subject may be identified.
        For example, an account may be identified by an opaque identifier, an
        email address, a phone number, a JWT "iss" claim and "sub" claim,
        etc., depending on the nature and needs of the transmitter and
        receiver.  Even within the context of a given transmitter and
        receiver relationship, it may be appropriate to identify different
        accounts in different ways, for example if some accounts only have
        email addresses associated with them while others only have phone
        numbers.  Therefore it can be necessary to indicate within a SET the
        mechanism by which a subject is being identified.

        To address this problem, this specification defines Subject
        Identifiers - JSON [RFC7519](https://datatracker.ietf.org/doc/html/rfc7519)
        objects containing information identifying a subject - and Identifier Formats -
        named sets of rules describing how to encode different kinds of subject
        identifying information (e.g., an email address, or an issuer and subject pair)
        as a Subject Identifier.
      discriminator:
        propertyName: format
      oneOf:
      - $ref: '#/components/schemas/SimpleSubject'
      - $ref: '#/components/schemas/Aliases'
      - $ref: '#/components/schemas/ComplexSubject'
      type: object
      x-one-of-name: Subject
    Error:
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
...
    ComplexSubject_application:
      allOf:
      - $ref: '#/components/schemas/SimpleSubject'
      - description: A Subject Identifier that identifies an application.
    ComplexSubject_device:
      allOf:
      - $ref: '#/components/schemas/SimpleSubject'
      - description: A Subject Identifier that identifies a device.
    ComplexSubject_group:
      allOf:
      - $ref: '#/components/schemas/SimpleSubject'
      - description: A Subject Identifier that identifies a group.
    ComplexSubject_org_unit:
      allOf:
      - $ref: '#/components/schemas/SimpleSubject'
      - description: A Subject Identifier that identifies an organizational unit.
    ComplexSubject_session:
      allOf:
      - $ref: '#/components/schemas/SimpleSubject'
      - description: A Subject Identifier that identifies a session.
    ComplexSubject_tenant:
      allOf:
      - $ref: '#/components/schemas/SimpleSubject'
      - description: A Subject Identifier that identifies a tenant.
...
  securitySchemes:
    BearerAuth:
      scheme: bearer
      type: http